server/requirements.txt
# Server Requirements
setuptools==80.9.0
flask==3.1.1
opencv-python==4.11.0.86
pillow==11.3.0
numpy==2.3.1
scikit-learn==1.7.0
requests==2.32.4
